+++ src/lib/constants/EnvVariables.ts
@@ -1,0 +1,3 @@
+export enum EnvVariablesConstants {
+  DEBUG_CP = 'DEBUG_CP',
+}